==> code/brms_model/code_from_xin/01_model.R <==
if (!require("install.load")) {
  install.packages("install.load")
  library(install.load)
}

# CRAN packages
list.of.packages <- c("tidyverse", "brms", "broom")

install.load::install_load(list.of.packages)

setwd(masstools::get_project_wd())
setwd("data_analysis/brms/")
rm(list = ls())
load("model-df.Rdata")
load("model.trans.Rdata")


model.df[,-c(1:8)] %>% 
  dplyr::select(-contains("days")) %>% 
  dplyr::select(-cluster) %>% 
  rowSums() %>% 
  plot()

###what is the cluster?
x <- model.trans %>%
  filter(cluster != 2) %>%
  mutate(cluster = fct_drop(cluster))

table(x$cluster)

# Model
# This should be modified to include all of the cytokines and genera of interest
# The format of the data frame should have each cytokine and genus as a column
cytokines <- c("IL17F")
genera <- c("Akkermansia")
################################

cytokine.model.output <- list()

# for (c in cytokines) {
#   genera.model.output <- list()
#   for (g in genera) {
    # f <- as.formula(paste(g, "~ days + (1|SubjectID) + cluster *", c,
    #                       sep = " "))
    # 
    # brm <-
    #   brms::brm(f,
    #             data = x,
    #             sparse = TRUE,
    #             family = negbinomial())

#     # Reformat output
#     # brm.df <- broom::tidy(brm)
#     brm.df <- broom.mixed::tidy(brm)
#     
#     # Create column to track the genus being modelled
#     brm.df$genus <- g
#     
#     # # Create list of outputs for each genus
#     # genera.model.output[[g]] <-
#     #   brm.df %>%
#     #   mutate(significant = !lower <= 0 & upper >= 0) %>%
#     #   filter(grepl(c, brm.df$term))
#     
#     genera.model.output[[g]] <-
#       brm.df %>%
#       mutate(significant = !conf.low <= 0 & conf.high >= 0) %>%
#       filter(grepl(c, brm.df$term))
#     
#   }
#   
#   cytokine.model.output[[c]] <- bind_rows(genera.model.output)
# }


##The hypothesis is that interaction term will be significant more frequently than the cytokine fixed term alone.
save(brm, file = "model-output.RData")
save(cytokine.model.output, file = "cytokine.model.output")
